--- casdoor_auth/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class CasdoorAuth(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "casdoor_auth"
