--- casdoor_auth/urls.py
@@ -1,8 +1,0 @@
-from django.urls import path
-
-from casdoor_auth import views
-
-urlpatterns = [
-    path('login/', views.toLogin, name='casdoor_sso'),
-    path('callback/', views.callback, name='callback'),
-]
