+++ setup.cfg
@@ -28,7 +28,9 @@ classifiers =
 
 [options]
 package_dir = 
-	=casdoor_auth
+	=src
+packages =
+	casdoor_auth
 setup_requires = 
 	setuptools
 install_requires =
