+++ src/casdoor_auth/__init__.py
@@ -1,0 +1,11 @@
+try:
+    from apps import *
+    from urls import *
+    from user import *
+    from views import *
+except ModuleNotFoundError or ImportError:
+    from .apps import *
+    from .urls import *
+    from .user import *
+    from .views import *
+
